save diary test ok

+++ src/presentation/mutate/response.rs
@@ -1,16 +1,16 @@
 use serde::Serialize;
 
-#[derive(Serialize, Debug)]
+#[derive(Serialize, Debug, Clone)]
 pub struct MutateResponse {
     pub result: MutateResult,
 }
 
-#[derive(Serialize, Debug)]
+#[derive(Serialize, Debug, Clone)]
 pub struct MutateResult {
     #[serde(rename = "rawContents")]
     pub raw_contents: Vec<String>,
     #[serde(rename = "mutatedText")]
     pub mutated_text: Vec<String>,
     #[serde(rename = "mutatedLength")]
-    pub mutated_length: Option<usize>,
+    pub mutated_length: Option<i32>,
 }
